Moved Exception to its own folder and Renamed DroneApiInterface to DroneApiService
